modified

+++ src/main/java/com/stackroute/services/TrackServiceImpl.java
@@ -18,28 +18,41 @@ import java.util.List;
 //@Profile({"trackService","default"})
 public class TrackServiceImpl implements TrackService
 {
-    @Autowired
     private TrackRepository trackRepository;
     @Autowired
-    public TrackServiceImpl(TrackRepository trackRepository) {
+    public TrackServiceImpl( TrackRepository trackRepository) {
         this.trackRepository = trackRepository;
     }
 
     @Override
-    public Track saveTrack(Track track)
+    public Track saveTrack(Track track) throws TrackAlreadyExistException {
+
+        Track savedTrack=trackRepository.save(track);
+        if(savedTrack==null)
+        {
+            throw new TrackAlreadyExistException("Track already exist in database");
+        }
+        return savedTrack;
+    }
+    @Override
+    public List<Track> getAllTrack() throws NullException
     {
-        return trackRepository.save(track);
+        List<Track> allTrack=trackRepository.findAll();
+        if(allTrack.size()==0)
+        {
+            throw new NullException();
+        }
+        return allTrack;
     }
 
     @Override
-    public List<Track> getAllTrack()
+    public boolean updateComment(int trackId,String comment) throws TrackNotFoundException
     {
-        return trackRepository.findAll();
-    }
 
-    @Override
-    public boolean updateComment(int trackId,String comment)
-    {
+        if(!trackRepository.existsById(trackId))
+        {
+            throw new TrackNotFoundException("Given trackId is not exist in database");
+        }
         Track trackToChangeCom=trackRepository.getOne(trackId);
         trackToChangeCom.setComments(comment);
         trackRepository.save(trackToChangeCom);
@@ -47,20 +60,35 @@ public class TrackServiceImpl implements TrackService
     }
 
     @Override
-    public void deleteTrack(int trackId)
+    public Track deleteTrack(int trackId) throws TrackNotFoundException
     {
+
+        if(!trackRepository.existsById(trackId))
+        {
+            throw new TrackNotFoundException("Given trackId is not exist in database");
+        }
         trackRepository.deleteById(trackId);
+        return trackRepository.getOne(trackId);
+
     }
 
     @Override
-    public List<Track> getTrackByTrackName(String trackName)
+    public List<Track> getTrackByTrackName(String trackName) throws NullException
     {
+        if(trackRepository.findAll().size()==0)
+        {
+            throw new NullException();
+        }
         return trackRepository.getTrackByTrackName(trackName);
     }
-    @Override
-    public boolean getTrackById(int trackId)
-    {
 
+    @Override
+    public boolean getTrackById(int trackId) throws NullException
+    {
+        if(!trackRepository.existsById(trackId))
+        {
+            throw new NullException();
+        }
         return trackRepository.existsById(trackId);
     }
 }
